requirements.txt
Django>=5.0
psycopg[binary]>=3.1
pgvector>=0.3.0
openai>=1.14.0
anthropic>=0.21.3
python-dotenv>=1.0.1
PyPDF2>=3.0.1
python-docx>=1.2.0
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0
requests>=2.31.0
cryptography>=41.0.0
django-redis>=5.4.0
redis>=5.0.0
google-generativeai>=0.3.0
cohere>=4.37
groq>=0.4.0
celery>=5.3.0
PyJWT>=2.8.0
gunicorn>=21.2.0
whitenoise>=6.6.0
